(feat) - Code updated to utilize Transport - The Module has been converted to utilize Puppet::Transport, the original class has been converted to a wrapper method, thus enabling backwards compatibility. - homehub.conf renamed to officehub.conf to better represent it’s function. - A philips hue emulator has been added to the module fixtures to aid in testing, with a default emulatorhub.conf file provided and a short description added to the README.md. - Functionality has been added in order to allow the port to be set. This is optional and if not set then the port will default to 80.
