+++ spec/fixtures/device.conf
@@ -1,3 +1,7 @@
-[homehub]
+[officehub]
 type philips_hue
-url  file:///Users/david.armstrong/projects/hue_rsapi/spec/fixtures/homehub.conf
+url  file:///Users/david.armstrong/projects/hue_rsapi/spec/fixtures/officehub.conf
+
+[emulatorhub]
+type philips_hue
+url  file:///Users/david.armstrong/projects/hue_rsapi/spec/fixtures/emulatorhub.conf
