+++ spec/fixtures/emulatorhub.conf
@@ -1,0 +1,3 @@
+host: 127.0.0.1
+key:  newdeveloper
+port: 80
