--- spec/fixtures/homehub.conf
@@ -1,6 +1,0 @@
-default: {
-  node: {
-    ip: 10.64.12.39
-    key: onmdTvd198bMrC6QYyVE9iasfYSeyAbAj3XyQzfL
-  }
-}
