+++ spec/fixtures/officehub.conf
@@ -1,0 +1,2 @@
+host: 10.64.12.39
+key: onmdTvd198bMrC6QYyVE9iasfYSeyAbAj3XyQzfL
